Try to fix bug on spot instances.
It seems the getusage type only works for eks or if we have karpenter tags. I expanded it to use the custom spot tag we setup on the configuration. this should work in any cluster scenario if the person configure the right tag. 

before this it was returning key with a trailing comma as described bellow: 
us-west-2,c5.4xlarge,linux,

Signed-off-by: Leonardo Rodrigues de Mello <lrodriguesdemello@zendesk.com>